Move STATE to odoo.constants
