--- src/modules/odoo/application/services/odoo.state.ts
@@ -1,6 +1,0 @@
-export enum STATE {
-  DRAFT = 'draft',
-  SALE = 'sale',
-  ASSIGNED = 'assigned',
-  DONE = 'done',
-}
