+++ src/modules/stellar/application/dto/create-order.dto.ts
@@ -1,6 +1,6 @@
 import { Equals } from 'class-validator';
 
-import { STATE } from '@/modules/odoo/application/services/odoo.state';
+import { STATE } from '@/modules/odoo/application/services/odoo.constants';
 
 import { SaleOrderDto } from './sale-order.dto';
