+++ src/modules/stellar/application/dto/deliver-order.dto.ts
@@ -1,6 +1,6 @@
 import { Equals } from 'class-validator';
 
-import { STATE } from '@/modules/odoo/application/services/odoo.state';
+import { STATE } from '@/modules/odoo/application/services/odoo.constants';
 
 import { StockPickingDto } from './stock-picking.dto';
